Move corpora.basic to corpora
