+++ skills_ml/algorithms/skill_extractors/base.py
@@ -5,7 +5,7 @@ from abc import ABCMeta, abstractmethod
 from collections import Counter
 
 
-from skills_ml.job_postings.corpora.basic import SimpleCorpusCreator
+from skills_ml.job_postings.corpora import SimpleCorpusCreator
 from skills_ml.algorithms.string_cleaners import NLPTransforms
 
 from typing import Dict, Callable, Text, Generator
