+++ skills_ml/job_postings/__init__.py
@@ -2,7 +2,7 @@ import json
 
 from descriptors import cachedproperty
 
-from skills_ml.job_postings.corpora.basic import SimpleCorpusCreator
+from skills_ml.job_postings.corpora import SimpleCorpusCreator
 
 
 class JobPosting(object):
