+++ skills_ml/job_postings/computed_properties/computers.py
@@ -8,7 +8,7 @@ from skills_ml.algorithms.string_cleaners import NLPTransforms
 from skills_ml.algorithms.jobtitle_cleaner.clean import JobTitleStringClean
 from skills_ml.algorithms.occupation_classifiers.classifiers import \
     SocClassifier
-from skills_ml.job_postings.corpora.basic import SimpleCorpusCreator
+from skills_ml.job_postings.corpora import SimpleCorpusCreator
 from skills_ml.job_postings.geography_queriers.cbsa_from_geocode import JobCBSAFromGeocodeQuerier
 from skills_ml.algorithms.skill_extractors import (
     ExactMatchSkillExtractor,
