+++ tests/test_skill_feature_creator.py
@@ -1,11 +1,11 @@
 from skills_ml.job_postings.common_schema import JobPostingCollectionSample
-from skills_ml.job_postings.corpora.basic import RawCorpusCreator
+from skills_ml.job_postings.corpora import RawCorpusCreator
 from skills_ml.algorithms.skill_feature_creator import SequenceFeatureCreator, EmbeddingFeature
 from skills_ml.algorithms.string_cleaners import NLPTransforms
 
 from skills_ml.algorithms.embedding.train import EmbeddingTrainer
 from skills_ml.algorithms.embedding.models import Word2VecModel, Doc2VecModel
-from skills_ml.job_postings.corpora.basic import Doc2VecGensimCorpusCreator, Word2VecGensimCorpusCreator
+from skills_ml.job_postings.corpora import Doc2VecGensimCorpusCreator, Word2VecGensimCorpusCreator
 
 import numpy as np
 from itertools import tee
